test(libeval): consolidate schema/messaging test pairs (spec 640)
Merge three schema/messaging siblings into their primaries to reduce Node
fork overhead:
- facilitator.test.js absorbs facilitator-messaging.test.js
- tee-writer.test.js absorbs tee-writer-schema.test.js
- trace-collector.test.js absorbs trace-collector-schema.test.js

Each pair shared imports and helper definitions; the merged file keeps
one set. Net: 3 fewer test-file forks per test run (each fork ~100ms of
Node startup on this machine).
